saluki_config::value

Enum Value

pub enum Value {
    String(Tag, String),
    Char(Tag, char),
    Bool(Tag, bool),
    Num(Tag, Num),
    Empty(Tag, Empty),
    Dict(Tag, BTreeMap<String, Value>),
    Array(Tag, Vec<Value>),
}

An enum representing all possible figment value variants.

Note that Value implements From<T> for all reasonable T:

use figment::value::Value;

let v = Value::from("hello");
assert_eq!(v.as_str(), Some("hello"));

Variants

String(Tag, String)

A string.

Char(Tag, char)

A character.

Bool(Tag, bool)

A boolean.

Num(Tag, Num)

A numeric value.

Empty(Tag, Empty)

A value with no value.

Dict(Tag, BTreeMap<String, Value>)

A dictionary: a map from String to Value.

Array(Tag, Vec<Value>)

A sequence/array/vector.

Implementations

impl Value

pub fn serialize<T>(value: T) -> Result<Value, Error>
where T: Serialize,

Serialize a Value from any T: Serialize.

use figment::value::{Value, Empty};

let value = Value::serialize(10i8).unwrap();
assert_eq!(value.to_i128(), Some(10));

let value = Value::serialize(()).unwrap();
assert_eq!(value, Empty::Unit.into());

let value = Value::serialize(vec![4, 5, 6]).unwrap();
assert_eq!(value, vec![4, 5, 6].into());

pub fn deserialize<'de, T>(&self) -> Result<T, Error>
where T: Deserialize<'de>,

Deserialize self into any deserializable T.

use figment::value::Value;

let value = Value::from("hello");
let string: String = value.deserialize().unwrap();
assert_eq!(string, "hello");

pub fn find(self, path: &str) -> Option<Value>

Looks up and returns the value at path path, where path is of the form a.b.c where a, b, and c are keys to dictionaries. If the key is empty, simply returns self. If the key is not empty and self or any of the values for non-leaf keys in the path are not dictionaries, returns None.

This method consumes self. See Value::find_ref() for a non-consuming variant.

Example

use figment::{value::Value, util::map};

let value = Value::from(map! {
    "apple" => map! {
        "bat" => map! {
            "pie" => 4usize,
        },
        "cake" => map! {
            "pumpkin" => 10usize,
        }
    }
});

assert!(value.clone().find("apple").is_some());
assert!(value.clone().find("apple.bat").is_some());
assert!(value.clone().find("apple.cake").is_some());

assert_eq!(value.clone().find("apple.bat.pie").unwrap().to_u128(), Some(4));
assert_eq!(value.clone().find("apple.cake.pumpkin").unwrap().to_u128(), Some(10));

assert!(value.clone().find("apple.pie").is_none());
assert!(value.clone().find("pineapple").is_none());

pub fn find_ref<'a>(&'a self, path: &str) -> Option<&'a Value>

Exactly like Value::find() but does not consume self, returning a reference to the found value, if any, instead.

Example

use figment::{value::Value, util::map};

let value = Value::from(map! {
    "apple" => map! {
        "bat" => map! {
            "pie" => 4usize,
        },
        "cake" => map! {
            "pumpkin" => 10usize,
        }
    }
});

assert!(value.find_ref("apple").is_some());
assert!(value.find_ref("apple.bat").is_some());
assert!(value.find_ref("apple.cake").is_some());

assert_eq!(value.find_ref("apple.bat.pie").unwrap().to_u128(), Some(4));
assert_eq!(value.find_ref("apple.cake.pumpkin").unwrap().to_u128(), Some(10));

assert!(value.find_ref("apple.pie").is_none());
assert!(value.find_ref("pineapple").is_none());

pub fn tag(&self) -> Tag

Returns the Tag applied to this value.

use figment::{Figment, Profile, value::Value, util::map};

let map: Value = Figment::from(("key", "value")).extract().unwrap();
let value = map.find_ref("key").expect("value");
assert_eq!(value.as_str(), Some("value"));
assert!(!value.tag().is_default());
assert_eq!(value.tag().profile(), Some(Profile::Global));

let map: Value = Figment::from(("key", map!["key2" => 123])).extract().unwrap();
let value = map.find_ref("key.key2").expect("value");
assert_eq!(value.to_i128(), Some(123));
assert!(!value.tag().is_default());
assert_eq!(value.tag().profile(), Some(Profile::Global));

pub fn as_str(&self) -> Option<&str>

Converts self into a &str if self is a Value::String.

Example

use figment::value::Value;

let value: Value = 123.into();
let converted = value.as_str();

pub fn into_string(self) -> Option<String>

Converts self into a String if self is a Value::String.

Example

use figment::value::Value;

let value: Value = 123.into();
let converted = value.into_string();

pub fn to_char(&self) -> Option<char>

Converts self into a char if self is a Value::Char.

Example

use figment::value::Value;

let value: Value = 123.into();
let converted = value.to_char();

pub fn to_bool(&self) -> Option<bool>

Converts self into a bool if self is a Value::Bool.

Example

use figment::value::Value;

let value: Value = 123.into();
let converted = value.to_bool();

pub fn to_num(&self) -> Option<Num>

Converts self into a Num if self is a Value::Num.

Example

use figment::value::Value;

let value: Value = 123.into();
let converted = value.to_num();

pub fn to_empty(&self) -> Option<Empty>

Converts self into a Empty if self is a Value::Empty.

Example

use figment::value::Value;

let value: Value = 123.into();
let converted = value.to_empty();

pub fn as_dict(&self) -> Option<&BTreeMap<String, Value>>

Converts self into a &Dict if self is a Value::Dict.

Example

use figment::value::Value;

let value: Value = 123.into();
let converted = value.as_dict();

pub fn into_dict(self) -> Option<BTreeMap<String, Value>>

Converts self into a Dict if self is a Value::Dict.

Example

use figment::value::Value;

let value: Value = 123.into();
let converted = value.into_dict();

pub fn as_array(&self) -> Option<&[Value]>

Converts self into a &[Value] if self is a Value::Array.

Example

use figment::value::Value;

let value: Value = 123.into();
let converted = value.as_array();

pub fn into_array(self) -> Option<Vec<Value>>

Converts self into a Vec<Value> if self is a Value::Array.

Example

use figment::value::Value;

let value: Value = 123.into();
let converted = value.into_array();

pub fn to_u128(&self) -> Option<u128>

Converts self into a u128 if self is an unsigned Value::Num variant.

Example

use figment::value::Value;

let value: Value = 123u8.into();
let converted = value.to_u128();
assert_eq!(converted, Some(123));

pub fn to_i128(&self) -> Option<i128>

Converts self into an i128 if self is an signed Value::Num variant.

Example

use figment::value::Value;

let value: Value = 123i8.into();
let converted = value.to_i128();
assert_eq!(converted, Some(123));

let value: Value = Value::from(5000i64);
assert_eq!(value.to_i128(), Some(5000i128));

pub fn to_f64(&self) -> Option<f64>

Converts self into an f64 if self is either a Num::F32 or Num::F64.

Example

use figment::value::Value;

let value: Value = 7.0f32.into();
let converted = value.to_f64();
assert_eq!(converted, Some(7.0f64));

let value: Value = Value::from(7.0f64);
assert_eq!(value.to_f64(), Some(7.0f64));

pub fn to_bool_lossy(&self) -> Option<bool>

Converts self to a bool if it is a Value::Bool, or if it is a Value::String or a Value::Num with a boolean interpretation.

The case-insensitive strings “true”, “yes”, “1”, and “on”, and the signed or unsigned integers 1 are interpreted as true.

The case-insensitive strings “false”, “no”, “0”, and “off”, and the signed or unsigned integers 0 are interpreted as false.

Example

use figment::value::Value;

let value = Value::from(true);
assert_eq!(value.to_bool_lossy(), Some(true));

let value = Value::from(1);
assert_eq!(value.to_bool_lossy(), Some(true));

let value = Value::from("YES");
assert_eq!(value.to_bool_lossy(), Some(true));

let value = Value::from(false);
assert_eq!(value.to_bool_lossy(), Some(false));

let value = Value::from(0);
assert_eq!(value.to_bool_lossy(), Some(false));

let value = Value::from("no");
assert_eq!(value.to_bool_lossy(), Some(false));

let value = Value::from("hello");
assert_eq!(value.to_bool_lossy(), None);

pub fn to_num_lossy(&self) -> Option<Num>

Converts self to a Num if it is a Value::Num or if it is a Value::String that parses as a usize (Num::USize), isize (Num::ISize), or f64 (Num::F64), in that order of precendence.

Examples

use figment::value::{Value, Num};

let value = Value::from(7_i32);
assert_eq!(value.to_num_lossy(), Some(Num::I32(7)));

let value = Value::from("7");
assert_eq!(value.to_num_lossy(), Some(Num::U8(7)));

let value = Value::from("-7000");
assert_eq!(value.to_num_lossy(), Some(Num::I16(-7000)));

let value = Value::from("7000.5");
assert_eq!(value.to_num_lossy(), Some(Num::F64(7000.5)));

pub fn to_actual(&self) -> Actual

Converts self into the corresponding Actual.

See also Num::to_actual() and Empty::to_actual(), which are called internally by this method.

Example

use figment::{value::Value, error::Actual};

assert_eq!(Value::from('a').to_actual(), Actual::Char('a'));
assert_eq!(Value::from(&[1, 2, 3]).to_actual(), Actual::Seq);

impl Value

This impl block contains no items.

Marker trait for “magic” values. Primarily for use with [Either].